Move webpack example as a top level example package. Otherwise lerna would not count it as part of the mono repo.
