+++ examples/parser/.gitignore
@@ -1,3 +1,2 @@
 /node_modules
 .idea/workspace.xml
-/webpack/lib
