+++ examples/parser/package.json
@@ -4,7 +4,6 @@
     "scripts": {
         "build": "npm-run-all build:minification build:webpack build:serialized-grammar",
         "build:minification": "grunt --gruntfile minification/gruntfile.js",
-        "build:webpack": "cd webpack && yarn install && yarn bundle && shx rm -rf node_modules && cd ..",
         "build:serialized-grammar": "node ./serialized_grammar/gen_serialized.js",
         "test": "mocha \"!(node_modules)/**/*spec.js\""
     },
